Allow overwriting an existing scheme with a new definition
Added a parameter allow_overwrite which defaults to False. If False, the behaviour of the save method on Scheme is the same as before, namely raising an Exception if there already exists a scheme of the same name. If set to True however upon saving an existing scheme of the same name will be overwritten and now exception will be raised.